Major refactoring of events and callbacks and their corresponding names.
Names are now [Event]Received for events and [Callback]Completed for callbacks.
Where [Event] and [Callback] are the corresponding names without the words
event or callback in the names.  This is more idiomatic .NET and better
indicates what the difference between an event and callback is.

All events and callbacks delegates are now stored in a backing dictionary
instead of in individual fields.  This is to help conserve memory, as only
those events and callbacks that are set will consume memory.

Future Notices:  As .NET 4.5 approaches, I am considering fully transitioning
the SDK to require it and thus getting rid of callback handlers entirely and
converting all blocking methods with callbacks into Async methods that follow
the new .NET 4.5 and C#/VB 5 async pattern.  This means you will no longer
need to set callbacks.  Just simply await the method.